1. split buildtime and runtime config 2. move runtime init from recorder to producer
