+++ quarkus-sa-token-main/quarkus-sa-token-resteasy/integration-tests/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-sa-token.exception-mapper-enabled=true
+sa-token.intercept.exception-mapper-enabled=true
