+++ quarkus-sa-token-main/quarkus-sa-token-resteasy/runtime/src/main/java/io/quarkiverse/satoken/core/config/SaRouteConfigForQuarkus.java
@@ -3,8 +3,9 @@ package io.quarkiverse.satoken.core.config;
 import java.util.List;
 import java.util.Optional;
 
-import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
+import io.quarkus.runtime.annotations.ConfigPhase;
+import io.quarkus.runtime.annotations.ConfigRoot;
 
 /**
  * SaRouteConfigForQuarkus
@@ -12,7 +13,7 @@ import io.quarkus.runtime.annotations.ConfigItem;
  * @author nayan
  * @date 2022/4/14 5:52 PM
  */
-@ConfigGroup
+@ConfigRoot(prefix = "sa-token", name = "route", phase = ConfigPhase.BUILD_AND_RUN_TIME_FIXED)
 public class SaRouteConfigForQuarkus {
     /**
      * enabled
