+++ quarkus-sa-token-plugins/quarkus-sa-token-dao-alone-redis/integration-tests/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-sa-token.exception-mapper-enabled=true
+sa-token.intercept.exception-mapper-enabled=true
 quarkus.application.name=sa-token-dao-alone-redis
 quarkus.redis.devservices.enabled=true
 quarkus.redis.devservices.port=63790
