+++ quarkus-sa-token-plugins/quarkus-sa-token-dao-redis-jackson/integration-tests/src/main/resources/application-annotation.properties
@@ -1,1 +1,1 @@
-sa-token.annotation-intercepted-enabled=true
+sa-token.intercept.annotation-intercepted-enabled=true
