+++ quarkus-sa-token-plugins/quarkus-sa-token-dao-redis-jackson/integration-tests/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-sa-token.exception-mapper-enabled=true
+sa-token.intercept.exception-mapper-enabled=true
 quarkus.application.name=sa-token-redis-jackson-test
 quarkus.redisson.single-server-config.address=redis://127.0.0.1:63790
 quarkus.redisson.single-server-config.database=0
